// src/helpers/Enums.ts
export enum SortFilterType {
  Text = "Text",
  Choice = "Choice",
  PartialChoice = "PartialChoice",
  Number = "Number",
  NumberRange = "NumberRange",
  Date = "Date",
  Time = "Time",
}

export enum SortDirection {
  Asc = "Asc",
  Desc = "Desc",
}

export enum DetailsListColumnType {
  Text = "Text",
  Date = "Date",
  Time = "Time",
  Image = "Image",
}